2/27 8:28 PM Errors with main (i havent tried out the subroutine)
I changed copy pasted all of the main loop all the way up to the debug initialize. I was really buggy and i changed the arrays. IDK if that fixed much. Also, Idk if you knew, but the debugger is only suppose to capture once you press the button, after you release it, and 6 times after you release it, so i added a counter in R11 which is set at six. Just make the Debug capture go straight to the POP if it equals zero, and add a SUB R11,#1 right before the POP